requirements.txt
google-cloud-bigquery==0.27.0
google-cloud-storage==1.1.1
